frm-spark-databricks-devops-dataops-dabs/02-ubear-eats-bundle-01-env/pyproject.toml
[project]
name = "ubear_eats_bundle_01_env"
version = "0.0.1"
authors = [{ name = "ubereats_data_engineer@gabrielpoc2025outlook.onmicrosoft.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    # Any dependencies for jobs and pipelines in this project can be added here
    # See also https://docs.databricks.com/dev-tools/bundles/library-dependencies
    #
    # LIMITATION: for pipelines, dependencies are cached during development;
    # add dependencies to the 'environment' section of your pipeline.yml file instead
]

[dependency-groups]
dev = [
    "pytest",
    "databricks-dlt",
    "databricks-connect>=15.4,<15.5",
]

[project.scripts]
main = "ubear_eats_bundle_01_env.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 125
